requirements.txt
discord.py
PyGithub
python-dotenv
PyYAML
aiofiles
claude-agent-sdk
